refactor: revise the document according to the OpenAPI specifications
