+++ src/user/responses/create-user-response.ts
@@ -3,7 +3,7 @@ import { ApiProperty } from "@nestjs/swagger";
 export class CreateUserResponse {
   @ApiProperty({
     description: "HTTP 回應代碼",
-    example: "201",
+    example: 201,
     type: "number",
   })
   public readonly statusCode: number;
@@ -11,7 +11,7 @@ export class CreateUserResponse {
   @ApiProperty({
     description: "創建成功回應",
     example: "創建成功",
-    type: "number",
+    type: "string",
   })
   public readonly message: string;
 }
